+++ vue-project/backend/public/index.html
@@ -1,1 +1,1 @@
-<!DOCTYPE html><html lang=en><head><meta charset=utf-8><title>openmate</title><script type=text/javascript src="//dapi.kakao.com/v2/maps/sdk.js?appkey=3036d01994527b266478b3224ee13ca9&libraries=services"></script><link href=https://unpkg.com/vuetify/dist/vuetify.min.css rel=stylesheet><link href=/static/css/app.f3862f174949839eecb624304bf3932c.css rel=stylesheet></head><body><div id=app></div><script src=/dist/build.js></script><script type=text/javascript src=/static/js/manifest.2ae2e69a05c33dfc65f8.js></script><script type=text/javascript src=/static/js/vendor.ade259903712307c5ac2.js></script><script type=text/javascript src=/static/js/app.720ead0fd2e025116cc4.js></script></body></html>
+<!DOCTYPE html><html lang=en><head><meta charset=utf-8><title>openmate</title><script type=text/javascript src="//dapi.kakao.com/v2/maps/sdk.js?appkey=3036d01994527b266478b3224ee13ca9&libraries=services"></script><link href=https://unpkg.com/vuetify/dist/vuetify.min.css rel=stylesheet><link href=/static/css/app.f3862f174949839eecb624304bf3932c.css rel=stylesheet></head><body><div id=app></div><script src=/dist/build.js></script><script type=text/javascript src=/static/js/manifest.2ae2e69a05c33dfc65f8.js></script><script type=text/javascript src=/static/js/vendor.ade259903712307c5ac2.js></script><script type=text/javascript src=/static/js/app.7d96c186c87dc8b206a9.js></script></body></html>
