+++ vue-project/frontend/src/config.js
@@ -1,0 +1,4 @@
+
+export default {
+    appKey: "3036d01994527b266478b3224ee13ca9"
+  }
